fix: add terms.ts
