+++ .github/workflows/production.yml
@@ -43,14 +43,14 @@ jobs:
       - name: Install dependencies
         run: pnpm install --frozen-lockfile
       
-      # # 5. 린트 검사
-      # - name: Run lint
-      #   run: pnpm run lint
-      #   continue-on-error: true
+      # 5. 린트 검사
+      - name: Run lint
+        run: pnpm run lint
+        continue-on-error: true
 
-      # # 6. 타입 체크
-      # - name: Type check
-      #   run: pnpm run type-check
+      # 6. 타입 체크
+      - name: Type check
+        run: pnpm run type-check
 
       # # 7. 테스트 실행
       # - name: Run tests
